MSEARCH-97: Remove DB system user service.
