+++ src/main/java/org/folio/search/configuration/properties/FolioEnvironment.java
@@ -2,12 +2,10 @@ package org.folio.search.configuration.properties;
 
 import lombok.Getter;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.stereotype.Component;
 
 @Getter
 @Component
-@ConditionalOnProperty("okapi.url")
 public class FolioEnvironment {
   private final String okapiUrl;
