--- src/test/java/org/folio/search/service/systemuser/DbSystemUserServiceTest.java
@@ -1,128 +1,0 @@
-package org.folio.search.service.systemuser;
-
-import static org.folio.search.model.service.ResultList.asSinglePage;
-import static org.hamcrest.MatcherAssert.assertThat;
-import static org.hamcrest.Matchers.is;
-import static org.hamcrest.Matchers.nullValue;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-
-import java.util.Optional;
-import org.folio.search.client.AuthnClient;
-import org.folio.search.client.PermissionsClient;
-import org.folio.search.client.UsersClient;
-import org.folio.search.configuration.properties.FolioSystemUserProperties;
-import org.folio.search.model.SystemUser;
-import org.folio.search.model.service.ResultList;
-import org.folio.search.repository.SystemUserRepository;
-import org.folio.search.service.context.FolioExecutionContextBuilder;
-import org.folio.search.utils.types.UnitTest;
-import org.folio.spring.FolioExecutionContext;
-import org.folio.spring.integration.XOkapiHeaders;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.ArgumentCaptor;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.http.ResponseEntity;
-
-@UnitTest
-@ExtendWith(MockitoExtension.class)
-class DbSystemUserServiceTest {
-  @Mock
-  private UsersClient usersClient;
-  @Mock
-  private AuthnClient authnClient;
-  @Mock
-  private PermissionsClient permissionsClient;
-  @Mock
-  private FolioExecutionContext executionContext;
-  @Mock
-  private FolioExecutionContextBuilder contextBuilder;
-  @Mock
-  private SystemUserRepository repository;
-
-  @Test
-  void shouldSaveUserWhenCreated() {
-    when(usersClient.query(any())).thenReturn(userNotExistResponse());
-
-    prepareSystemUser(systemUser());
-
-    verify(usersClient).saveUser(any());
-    verify(permissionsClient).assignPermissionsToUser(any());
-
-    var captor = ArgumentCaptor.forClass(SystemUser.class);
-    verify(repository).save(captor.capture());
-
-    assertThat(captor.getValue().getUsername(), is(systemUser().getUsername()));
-    assertThat(captor.getValue().getToken(), nullValue());
-  }
-
-  @Test
-  void shouldSaveUserWhenUpdated() {
-    when(usersClient.query(any())).thenReturn(userExistsResponse());
-    when(permissionsClient.getUserPermissions(any())).thenReturn(ResultList.empty());
-
-    prepareSystemUser(systemUser());
-
-    verify(permissionsClient, times(2)).addPermission(any(), any());
-
-    var captor = ArgumentCaptor.forClass(SystemUser.class);
-    verify(repository).save(captor.capture());
-
-    assertThat(captor.getValue().getUsername(), is(systemUser().getUsername()));
-    assertThat(captor.getValue().getToken(), nullValue());
-  }
-
-  @Test
-  void getSystemUser_shouldLogInUserWhenNoToken() {
-    when(repository.findOneByUsername(any())).thenReturn(Optional.of(new SystemUser()));
-    when(authnClient.getApiKey(any())).thenReturn(ResponseEntity.status(200)
-      .header(XOkapiHeaders.TOKEN, "token").build());
-
-    var user = systemUserService(systemUser()).getSystemUser("tenant");
-
-    assertThat(user.getToken(), is("token"));
-    verify(authnClient, times(1)).getApiKey(any());
-  }
-
-  @Test
-  void shouldThrowExceptionIfNoUser() {
-    when(repository.findOneByUsername(any())).thenReturn(Optional.empty());
-
-    var systemUserService = systemUserService(systemUser());
-
-    assertThrows(IllegalStateException.class,
-      () -> systemUserService.getSystemUser("tenant"));
-  }
-
-  private FolioSystemUserProperties systemUser() {
-    return FolioSystemUserProperties.builder()
-      .password("password")
-      .username("username")
-      .permissionsFilePath("permissions/test-permissions.csv")
-      .build();
-  }
-
-  private ResultList<UsersClient.User> userExistsResponse() {
-    return asSinglePage(new UsersClient.User());
-  }
-
-  private ResultList<UsersClient.User> userNotExistResponse() {
-    return ResultList.empty();
-  }
-
-  private SystemUserService systemUserService(FolioSystemUserProperties properties) {
-    var prepareSystemUserService = new PrepareSystemUserService(permissionsClient,
-      usersClient, authnClient, contextBuilder, properties);
-    return new DbSystemUserService(prepareSystemUserService,
-      repository, contextBuilder, properties);
-  }
-
-  private void prepareSystemUser(FolioSystemUserProperties properties) {
-    systemUserService(properties).prepareSystemUser(executionContext);
-  }
-}
